Did restructuring and made repo validation interceptor an optional bean instead as it makes it more clean
