+++ src/main/java/ca/uhn/fhir/jpa/starter/common/FhirServerConfigR5.java
@@ -1,6 +1,7 @@
-package ca.uhn.fhir.jpa.starter;
+package ca.uhn.fhir.jpa.starter.common;
 
 import ca.uhn.fhir.jpa.config.r5.JpaR5Config;
+import ca.uhn.fhir.jpa.starter.ElasticsearchConfig;
 import ca.uhn.fhir.jpa.starter.annotations.OnR5Condition;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
